requirements.txt
Django==1.11.9
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
psycopg2==2.7.3.2